requirements.txt
streamlit==1.32.0
openai-whisper==20231117
python-dotenv==1.0.0
python-dateutil==2.8.2
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-magic==0.4.27
dateparser==1.1.8
boto3==1.34.0
python-magic-bin==0.4.14; sys_platform == 'win32'
